+++ clustercontroller/clustercontroller.py
@@ -12,6 +12,7 @@ import uuid
 from pathlib import Path
 from time import sleep, time
 
+from func_timeout import func_timeout, FunctionTimedOut
 from raven import Client
 from raven.handlers.logging import SentryHandler
 
@@ -241,7 +242,13 @@ class ClusterController:
     # Environment settings
     environment = os.environ.get('ENVIRONMENT')
     service = os.environ.get('SERVICE')
-    etcd_host = os.environ.get('ETCD_HOST')
+
+    if os.environ.get('ETCD_HOSTS') is not None:
+        if ',' in os.environ.get('ETCD_HOSTS'):
+            etcd_hosts = os.environ.get('ETCD_HOSTS').split(',')
+        else:
+            etcd_hosts = list(os.environ.get('ETCD_HOSTS'))
+
     etcd_port = os.environ.get('ETCD_PORT')
 
     try:
@@ -300,8 +307,8 @@ class ClusterController:
                 self.logger.error(f'ETCD Port should be a valid integer value.', extra={'stack': True, })
                 self.state = 'stopping'
 
-        if not isinstance(self.etcd_host, str) or not isinstance(self.etcd_port, int):
-            self.logger.error(f'No valid ETCD host and/or port specified: {self.etcd_host}:{self.etcd_port}',
+        if not isinstance(self.etcd_hosts, list) or not isinstance(self.etcd_port, int):
+            self.logger.error(f'No valid ETCD hosts and/or port specified: {self.etcd_hosts}:{self.etcd_port}',
                               extra={'stack': True, })
             self.state = 'stopping'
 
@@ -316,19 +323,25 @@ class ClusterController:
         connected = False
 
         timeout = time() + 30
-        attempt = 0
-        while not connected and time() < timeout:
-            attempt += 1
-            self.logger.info(f'Trying to connect to ETCD, attempt: {attempt}', extra={'stack': True, })
-            try:
-                self.etcd_client = etcd.Client(host=self.etcd_host, port=int(self.etcd_port), allow_reconnect=True)
-                machines = self.etcd_client.machines
-                if len(machines) >= 1:
-                    connected = True
-                    self.logger.info(f'Connected to ETCD machines: {machines}', extra={'stack': True, })
-            except etcd.EtcdException as error:
-                self.logger.info(f'Unable to connect to ETCD: {error}', extra={'stack': True, })
-            sleep(1)
+
+        while time() < timeout and not connected:
+            for host in self.etcd_hosts:
+                self.logger.info(f'Trying to connect to ETCD host {host}', extra={'stack': True, })
+                try:
+                    self.etcd_client = func_timeout(5, etcd.Client, args=(), kwargs={'host': host,
+                                                                                     'port': int(self.etcd_port),
+                                                                                     'allow_reconnect': True})
+                    machines = self.etcd_client.machines
+                    if len(machines) >= 1:
+                        connected = True
+                        self.logger.info(f'Connected to ETCD machines: {machines}', extra={'stack': True, })
+                        break
+                except FunctionTimedOut as error:
+                    self.logger.info(f'Timeout while connecting to ETCD host {host}', extra={'stack': True, })
+                except etcd.EtcdException as error:
+                    self.logger.info(f'Unable to connect to ETCD: {error}', extra={'stack': True, })
+                sleep(1)
+
 
         if not connected:
             self.logger.warning(f'Unable to connect to ETCD, giving up...', extra={'stack': True, })
@@ -477,6 +490,12 @@ class ClusterController:
                     self.logger.info('Member removed from ETCD.', extra={'stack': True, })
             except etcd.EtcdException:
                 pass
+
+            # Handle removal of filesystem locks
+            if self.filesystem_locks:
+                for folder in self.filesystem_locks:
+                    lockfile_path = Path(str(folder).rstrip('/') + '/lock')
+                    os.remove(lockfile_path)
 
             self.terminate_schedule_event.set()
             self.terminate_run_event.set()
@@ -776,7 +795,7 @@ class ClusterController:
                 lock_timestamp = int(lock_content.split(':')[1])
 
                 if lock_container == self.container and lock_timestamp + 10000 <= current_timestamp or \
-                        lock_content != self.container and lock_timestamp + 30000 < current_timestamp:
+                        lock_container != self.container and lock_timestamp + 30000 < current_timestamp:
                     write_new_lockfile = True
         else:
             write_new_lockfile = True