+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+
+## 0.1.12 - 2019-04-3
+
+* Bugfix
+
+## 0.1.11 - 2019-04-3
+
+* Feature: Changed ETCD_HOST to ETCD_HOSTS which now allows for a list of ETCD servers to be specified (comma seperated)
+
+## 0.1.10 - 2019-04-3
+
+* Feature: Remove lockfile when cluster controller exits
+
 ## 0.1.9 - 2019-04-3
 
 * Bug: get_filesystem_lock didn't return True when no lock file existed